separate classes and tests
