+++ haystack/preview/components/retrievers/__init__.py
@@ -1,3 +1,4 @@
-from haystack.preview.components.retrievers.memory import MemoryBM25Retriever, MemoryEmbeddingRetriever
+from haystack.preview.components.retrievers.memory_bm25_retriever import MemoryBM25Retriever
+from haystack.preview.components.retrievers.memory_embedding_retriever import MemoryEmbeddingRetriever
 
 __all__ = ["MemoryBM25Retriever", "MemoryEmbeddingRetriever"]
